+++ app/controllers/admin_users/passwords_controller.rb
@@ -13,7 +13,7 @@ class AdminUsers::PasswordsController < Devise::PasswordsController
   #   super
   # end
 
-  # GET /admin/password/edit パスワード再設定
+  # GET /admin/password パスワード再設定
   def edit
     return redirect_to new_admin_user_password_path, alert: invalid_token_message unless valid_reset_password_token?(params[:reset_password_token])