+++ app/controllers/users/auth/token_validations_controller.rb
@@ -4,6 +4,7 @@ class Users::Auth::TokenValidationsController < DeviseTokenAuth::TokenValidation
   include DeviseTokenAuth::Concerns::SetUserByToken
   skip_before_action :verify_authenticity_token
   prepend_before_action :not_acceptable_response_not_api_accept
+  prepend_before_action :update_request_uid_header
 
   # GET /users/auth/validate_token(.json) トークン検証API(処理)
   # def validate_token