+++ spec/requests/users/auth/confirmations_spec.rb
@@ -216,7 +216,9 @@ RSpec.describe 'Users::Auth::Confirmations', type: :request do
   #   確認日時: ない（未確認）, 確認送信日時より前（未確認）, 確認送信日時より後（確認済み）
   #   ＋リダイレクトURL: ある, ない, ホワイトリストにない
   describe 'GET #show' do
-    subject { get user_auth_confirmation_path(confirmation_token: confirmation_token, redirect_url: @redirect_url), headers: auth_headers.merge(accept_headers) }
+    subject do
+      get user_auth_confirmation_path(confirmation_token: confirmation_token, redirect_url: @redirect_url), headers: auth_headers.merge(accept_headers)
+    end
     let(:current_user) { User.find(send_user.id) }
 
     # テスト内容